Create ThreatHuntAndCompare.sh
This script saves base (hopefully uninfected) files to locations and allows the user to compare when needed to see if they've been modified. This should be modified to be used with a cronjob or systemd timer to periodically check for modified or infected files. There seems to be some unintended things going on with the cmp commands. I do not believe they effect the script ability to perform but it should be tested all the same. change the last few lines if you'd like to add the cronjob and remove the base script. I'm going to start working on a windows version.